algo_production/requirements.txt
easyocr==1.7.2
fastapi==0.115.4
matplotlib==3.9.2
numpy==2.1.3
openai==0.28.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
roboflow==1.1.49
skimage==0.0
torch==2.5.1
torchvision==0.20.1
transformers==4.46.2
ultralytics==8.3.27
uvicorn==0.32.0
